+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - CLI:
   - `--variant` to specify which variants will be generated (See #737)
+  - `--defs-from-project` allows using KiCad variables for the preprocessor
+    (See #643)
 - Variants and filters:
   - Outputs with variants can now specify an `exclude_filter` in addition to the
     variant. Like it was possible for transform and dnf